requirements.txt
enum==0.4.6
nose==1.3.7
Pillow==3.3.0
